+++ lib/core/src/display/render/webgl/glsl.rs
@@ -8,9 +8,10 @@ use code_builder::{CodeBuilder, HasCodeRepr};
 use shapely::derive_clone_plus;
 
 
-// =============================================================================
-// === Expr ====================================================================
-// =============================================================================
+
+// =================================================================================================
+// === Expr ========================================================================================
+// =================================================================================================
 
 /// Any GLSL expression, like function call, or math operations.
 #[derive(Shrinkwrap,Clone,Debug)]
@@ -33,6 +34,7 @@ impl From<&String> for Expr {
         Expr::new(t)
     }
 }
+
 
 // === ExprUnboxed ===
 
@@ -72,6 +74,7 @@ impl From<&String> for ExprUnboxed {
 }
 
 
+
 // ===============
 // === RawCode ===
 // ===============
@@ -93,6 +96,7 @@ impl HasCodeRepr for RawCode {
         builder.write(&self.str)
     }
 }
+
 
 
 // ==================
@@ -128,6 +132,7 @@ impl From<&str> for Identifier {
 }
 
 
+
 // =============
 // === Block ===
 // =============
@@ -153,6 +158,7 @@ impl HasCodeRepr for Block {
         }
     }
 }
+
 
 
 // ==================
@@ -183,9 +189,9 @@ impl HasCodeRepr for Assignment {
 
 
 
-// =============================================================================
-// === Statement ===============================================================
-// =============================================================================
+// =================================================================================================
+// === Statement ===================================================================================
+// =================================================================================================
 
 /// Top-level statement, like function declaration.
 #[derive(Clone,Debug)]
@@ -210,6 +216,7 @@ impl From<PrecisionDecl> for Statement {
         Self::PrecisionDecl(t)
     }
 }
+
 
 
 // ================
@@ -241,6 +248,7 @@ impl<T:Into<Expr>> Add<T> for Function {
         self.body.add(t)
     }
 }
+
 
 
 // =====================
@@ -284,9 +292,9 @@ impl HasCodeRepr for PrecisionDecl {
 
 
 
-// =============================================================================
-// === AST Elements ============================================================
-// =============================================================================
+// =================================================================================================
+// === AST Elements ================================================================================
+// =================================================================================================
 
 
 // ============
@@ -314,6 +322,7 @@ impl HasCodeRepr for Type {
 }
 
 derive_clone_plus!(Type);
+
 
 
 // ================
@@ -541,9 +550,9 @@ impl From<&Precision> for Precision {
 
 
 
-// =============================================================================
-// === Module ==================================================================
-// =============================================================================
+// =================================================================================================
+// === Module ======================================================================================
+// =================================================================================================
 
 /// Translation unit definition. It represents the whole GLSL file.
 #[derive(Clone,Debug)]
@@ -622,4 +631,3 @@ impl HasCodeRepr for Module {
         builder.add(&self.main);
     }
 }
-
