--- lib/core/src/display/symbol/geometry/primitive/mesh/buffer/data.rs
@@ -1,4 +1,0 @@
-use crate::data::seq::observable::Observable;
-
-/// Vector with attached callbacks listening for changes.
-pub type Data<T,OnSet,OnResize> = Observable<Vec<T>,OnSet,OnResize>;
