+++ lib/core/src/display/symbol/material.rs
@@ -4,6 +4,7 @@ use crate::prelude::*;
 
 use crate::display::render::webgl::Context;
 use crate::display::render::webgl;
+use crate::display::render::webgl::glsl;
 use crate::data::function::callback::*;
 use crate::data::dirty;
 use crate::data::dirty::traits::*;
@@ -58,21 +59,21 @@ impl<OnDirty: Callback0> Material<OnDirty> {
 
                 // FIXME: Hardcoded variables until we get proper shaders EDSL.
 
-                let mut shader_cfg     = shader::builder::ShaderConfig::new();
-                let mut shader_builder = shader::builder::ShaderBuilder::new();
-                shader_cfg.insert_attribute        ("bbox"            , shader::glsl::PrimType::Vec2);
-                shader_cfg.insert_attribute        ("uv"              , shader::glsl::PrimType::Vec2);
-                shader_cfg.insert_attribute        ("transform"       , shader::glsl::PrimType::Mat4);
-                shader_cfg.insert_shared_attribute ("local"           , shader::glsl::PrimType::Vec3);
-                shader_cfg.insert_uniform          ("view_projection" , shader::glsl::PrimType::Mat4);
-                shader_cfg.insert_output           ("color"           , shader::glsl::PrimType::Vec4);
+                let mut shader_cfg     = shader::ShaderConfig::new();
+                let mut shader_builder = shader::ShaderBuilder::new();
+                shader_cfg.insert_attribute        ("bbox"            , glsl::PrimType::Vec2);
+                shader_cfg.insert_attribute        ("uv"              , glsl::PrimType::Vec2);
+                shader_cfg.insert_attribute        ("transform"       , glsl::PrimType::Mat4);
+                shader_cfg.insert_shared_attribute ("local"           , glsl::PrimType::Vec3);
+                shader_cfg.insert_uniform          ("view_projection" , glsl::PrimType::Mat4);
+                shader_cfg.insert_output           ("color"           , glsl::PrimType::Vec4);
 
-                let vtx_template = shader::builder::CodeTemplete::from_main("
+                let vtx_template = shader::CodeTemplete::from_main("
                 mat4 model_view_projection = view_projection * transform;
                 local                      = vec3((uv - 0.5) * bbox, 0.0);
                 gl_Position                = model_view_projection * vec4(local,1.0);
                 ");
-                let frag_template = shader::builder::CodeTemplete::from_main("
+                let frag_template = shader::CodeTemplete::from_main("
                 out_color = vec4(1.0,1.0,1.0,1.0);
                 ");
                 shader_builder.compute(&shader_cfg,vtx_template,frag_template);
